+++ ucm/transport/kv/asu/trans_provider/aiv_provider/src/ub/oob/tcp_client.h
@@ -61,7 +61,7 @@ public:
         uint32_t connectTimeoutMs{3000};
         uint32_t ioTimeoutMs{3000};
         TcpClientLocalEndpoint local;
-        uint32_t kato{5000};
+        uint16_t kato{5000};
         uint16_t cmdDepth{1024};
     };
